+++ src/pyedb/dotnet/edb_core/edb_data/hfss_pi_simulation_setup_data.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 from pyedb.dotnet.edb_core.general import convert_py_list_to_net_list
-from pyedb.dotnet.edb_core.sim_setup_data.data.simulation_setup import SimulationSetup
+from pyedb.dotnet.edb_core.utilities.simulation_setup import SimulationSetup
 from pyedb.dotnet.edb_core.sim_setup_data.data.sweep_data import SweepData
 from pyedb.generic.data_handlers import pyedb_function_handler
 from pyedb.generic.general_methods import generate_unique_name
